# app/routers/staff.py
"""
Staff / Admin management router

GET    /staff/pending               -> list unapproved accounts (site admin only)
GET    /staff/                      -> list all staff accounts (site admin only)
PATCH  /staff/{id}/approve          -> approve or revoke a staff account (site admin only)
PATCH  /staff/{id}/disable          -> disable a staff account (site admin only)
PATCH  /staff/{id}/enable           -> re-enable a disabled staff account (site admin only)
PATCH  /staff/{id}/role             -> promote / demote role (site admin only)
POST   /staff/{id}/recovery-code    -> generate a 6-digit recovery code (site admin only)
"""

import secrets
from datetime import UTC, datetime
from typing import Annotated

from fastapi import APIRouter, Body, Depends, HTTPException, status
from sqlmodel import Session, col, select

from app.core.database import get_session
from app.core.deps import get_site_admin
from app.models.models import PossibleRoles, Role, User
from app.openapi_examples.staff_examples import (
    APPROVE_REQUEST_EXAMPLES,
    APPROVE_RESPONSES,
    DISABLE_REQUEST_EXAMPLES,
    DISABLE_RESPONSES,
    ENABLE_RESPONSES,
    LIST_PENDING_RESPONSES,
    LIST_STAFF_RESPONSES,
    ROLE_REQUEST_EXAMPLES,
    ROLE_RESPONSES,
)
from app.schemas import ApproveUserRequest, DisableUserRequest, RecoveryCodeResponse, RoleRequest, UserPublic

router = APIRouter(prefix="/staff", tags=["staff"])


# -- Helpers ------------------------------------------------------------------


def _get_staff_or_404(user_id: int, session: Session) -> User:
    user = session.get(User, user_id)
    if not user:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Staff user not found.")
    return user


def _get_role_or_404(role: PossibleRoles, session: Session) -> Role:
    row = session.exec(select(Role).where(Role.name == role.value)).first()
    if not row:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Role '{role.value}' not found in database. Run seed.",
        )
    return row


# -- List all staff -----------------------------------------------------------


@router.get(
    "/",
    response_model=list[UserPublic],
    summary="List all staff accounts (site admin only)",
    responses=LIST_STAFF_RESPONSES,
)
async def list_staff(
    session: Session = Depends(get_session),
    _admin: User = Depends(get_site_admin),
) -> list[User]:
    return list(session.exec(select(User).order_by(col(User.created_at))).all())


# -- List pending approvals ---------------------------------------------------


@router.get(
    "/pending",
    response_model=list[UserPublic],
    summary="List unapproved staff accounts (site admin only)",
    responses=LIST_PENDING_RESPONSES,
)
async def list_pending(
    session: Session = Depends(get_session),
    _admin: User = Depends(get_site_admin),
) -> list[User]:
    return list(
        session.exec(
            select(User)
            .where(User.is_approved == False)  # noqa: E712
            .order_by(col(User.created_at))
        ).all()
    )


# -- Approve / revoke ---------------------------------------------------------


@router.patch(
    "/{user_id}/approve",
    response_model=UserPublic,
    summary="Approve or revoke a staff account (site admin only)",
    responses=APPROVE_RESPONSES,
)
async def set_approval(
    user_id: int,
    body: Annotated[ApproveUserRequest, Body(openapi_examples=APPROVE_REQUEST_EXAMPLES)],
    session: Session = Depends(get_session),
    admin: User = Depends(get_site_admin),
) -> User:
    user = _get_staff_or_404(user_id, session)

    if user.id == admin.id and not body.is_approved:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Site admins cannot revoke their own approval.",
        )

    user.is_approved = body.is_approved
    session.add(user)
    session.commit()
    session.refresh(user)
    return user


# -- Change role --------------------------------------------------------------


@router.patch(
    "/{user_id}/role",
    response_model=UserPublic,
    summary="Promote or demote a staff member's role (site admin only)",
    responses=ROLE_RESPONSES,
)
async def set_role(
    user_id: int,
    body: Annotated[RoleRequest, Body(openapi_examples=ROLE_REQUEST_EXAMPLES)],
    session: Session = Depends(get_session),
    admin: User = Depends(get_site_admin),
) -> User:
    user = _get_staff_or_404(user_id, session)

    if user.id == admin.id and body.role != PossibleRoles.SITE_ADMIN:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Site admins cannot demote themselves.",
        )

    role_row = _get_role_or_404(body.role, session)
    if role_row.id is None:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="Role row has no ID. This is a data integrity error.",
        )
    user.role_id = role_row.id
    session.add(user)
    session.commit()
    session.refresh(user)
    return user


# -- Disable staff user -------------------------------------------------------


@router.patch(
    "/{user_id}/disable",
    response_model=UserPublic,
    summary="Disable a staff account (site admin only)",
    responses=DISABLE_RESPONSES,
)
async def disable_staff(
    user_id: int,
    body: Annotated[DisableUserRequest, Body(openapi_examples=DISABLE_REQUEST_EXAMPLES)],
    session: Session = Depends(get_session),
    admin: User = Depends(get_site_admin),
) -> User:
    """
    Disable a staff account (soft-delete).

    Disabled accounts cannot log in but their data is preserved for
    audit trails. Records who disabled the account, when, and why.
    """
    user = _get_staff_or_404(user_id, session)

    if user.id == admin.id:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Site admins cannot disable their own account.",
        )

    if user.is_disabled:
        raise HTTPException(
            status_code=status.HTTP_409_CONFLICT,
            detail="This account is already disabled.",
        )

    user.is_disabled = True
    user.disabled_at = datetime.now(UTC)
    user.disabled_by_id = admin.id
    user.disabled_reason = body.reason
    session.add(user)
    session.commit()
    session.refresh(user)
    return user


# -- Enable staff user --------------------------------------------------------


@router.patch(
    "/{user_id}/enable",
    response_model=UserPublic,
    summary="Re-enable a disabled staff account (site admin only)",
    responses=ENABLE_RESPONSES,
)
async def enable_staff(
    user_id: int,
    session: Session = Depends(get_session),
    _admin: User = Depends(get_site_admin),
) -> User:
    """
    Re-enable a previously disabled staff account.

    Clears the disable fields, allowing the user to log in again.
    """
    user = _get_staff_or_404(user_id, session)

    if not user.is_disabled:
        raise HTTPException(
            status_code=status.HTTP_409_CONFLICT,
            detail="This account is not disabled.",
        )

    user.is_disabled = False
    user.disabled_at = None
    user.disabled_by_id = None
    user.disabled_reason = None
    session.add(user)
    session.commit()
    session.refresh(user)
    return user


# ── Generate recovery code ────────────────────────────────────────────────────


@router.post(
    "/{user_id}/recovery-code",
    response_model=RecoveryCodeResponse,
    summary="Generate a 6-digit password recovery code for a staff account (admin only)",
    status_code=status.HTTP_200_OK,
)
async def generate_recovery_code(
    user_id: int,
    session: Session = Depends(get_session),
    _admin: User = Depends(get_site_admin),
) -> RecoveryCodeResponse:
    """
    Generate a one-time 6-digit recovery code for the given user.

    The code expires after 72 hours and is invalidated once used.
    If the user already had a code, it is overwritten (implicitly invalidated).
    The admin is responsible for communicating the code to the user out-of-band.
    """
    user = _get_staff_or_404(user_id, session)

    code = f"{secrets.randbelow(1_000_000):06d}"
    user.recovery_code = code
    user.recovery_code_created_at = datetime.now(UTC)
    user.recovery_code_used = False
    session.add(user)
    session.commit()

    return RecoveryCodeResponse(username=user.username, code=code)