# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alter Event Tracker",
    "version": "0.1.0"
  },
  "paths": {
...
    "/staff/": {
      "get": {
        "tags": [
          "staff"
        ],
        "summary": "List all staff accounts (site admin only)",
        "operationId": "list_staff_staff__get",
        "responses": {
          "200": {
            "description": "All staff accounts",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserPublic"
                  },
                  "type": "array",
                  "title": "Response List Staff Staff  Get"
                },
                "examples": {
                  "all_staff": {
                    "summary": "Mix of approved, pending, and disabled",
                    "value": [
                      {
                        "id": 1,
                        "username": "maria_admin",
                        "is_approved": true,
                        "role": {
                          "id": 2,
                          "name": "SITE_ADMIN"
                        },
                        "created_at": "2024-03-15T20:00:00Z",
                        "is_disabled": false
                      },
                      {
                        "id": 2,
                        "username": "carlos_staff",
                        "is_approved": true,
                        "role": {
                          "id": 1,
                          "name": "USER"
                        },
                        "created_at": "2024-03-20T18:30:00Z",
                        "is_disabled": false
                      },
                      {
                        "id": 3,
                        "username": "nuevo_voluntario",
                        "is_approved": false,
                        "role": {
                          "id": 1,
                          "name": "USER"
                        },
                        "created_at": "2024-03-22T21:15:00Z",
                        "is_disabled": false
                      },
                      {
                        "id": 4,
                        "username": "ex_voluntario",
                        "is_approved": true,
                        "role": {
                          "id": 1,
                          "name": "USER"
                        },
                        "created_at": "2024-03-15T20:00:00Z",
                        "is_disabled": true,
                        "disabled_at": "2024-03-22T21:15:00Z",
                        "disabled_reason": "Left the organization"
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "examples": {
                  "missing_token": {
                    "summary": "No token provided",
                    "value": {
                      "detail": "Not authenticated"
                    }
                  },
                  "invalid_token": {
                    "summary": "Invalid or expired JWT",
                    "value": {
                      "detail": "Could not validate credentials"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "examples": {
                  "admin_required": {
                    "summary": "Org admin privileges needed",
                    "value": {
                      "detail": "Organization admin privileges required."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/staff/pending": {
      "get": {
        "tags": [
          "staff"
        ],
        "summary": "List unapproved staff accounts (site admin only)",
        "operationId": "list_pending_staff_pending_get",
        "responses": {
          "200": {
            "description": "Unapproved staff accounts",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserPublic"
                  },
                  "type": "array",
                  "title": "Response List Pending Staff Pending Get"
                },
                "examples": {
                  "pending_users": {
                    "summary": "Staff awaiting approval",
                    "value": [
                      {
                        "id": 3,
                        "username": "nuevo_voluntario",
                        "is_approved": false,
                        "role": {
                          "id": 1,
                          "name": "USER"
                        },
                        "created_at": "2024-03-22T21:15:00Z",
                        "is_disabled": false
                      }
                    ]
                  },
                  "no_pending": {
                    "summary": "No pending approvals",
                    "value": []
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "examples": {
                  "missing_token": {
                    "summary": "No token provided",
                    "value": {
                      "detail": "Not authenticated"
                    }
                  },
                  "invalid_token": {
                    "summary": "Invalid or expired JWT",
                    "value": {
                      "detail": "Could not validate credentials"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "examples": {
                  "admin_required": {
                    "summary": "Org admin privileges needed",
                    "value": {
                      "detail": "Organization admin privileges required."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/staff/{user_id}/approve": {
      "patch": {
        "tags": [
          "staff"
        ],
        "summary": "Approve or revoke a staff account (site admin only)",
        "operationId": "set_approval_staff__user_id__approve_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveUserRequest"
              },
              "examples": {
                "approve": {
                  "summary": "Approve a staff account",
                  "description": "Set is_approved to true so the user can log in",
                  "value": {
                    "is_approved": true
                  }
                },
                "revoke": {
                  "summary": "Revoke approval",
                  "description": "Set is_approved to false to lock out the user",
                  "value": {
                    "is_approved": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Approval status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                },
                "examples": {
                  "approved": {
                    "summary": "User now approved",
                    "value": {
                      "id": 3,
                      "username": "nuevo_voluntario",
                      "is_approved": true,
                      "role": {
                        "id": 1,
                        "name": "USER"
                      },
                      "created_at": "2024-03-22T21:15:00Z",
                      "is_disabled": false
                    }
                  },
                  "revoked": {
                    "summary": "User approval revoked",
                    "value": {
                      "id": 2,
                      "username": "carlos_staff",
                      "is_approved": false,
                      "role": {
                        "id": 1,
                        "name": "USER"
                      },
                      "created_at": "2024-03-20T18:30:00Z",
                      "is_disabled": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid self-operation",
            "content": {
              "application/json": {
                "examples": {
                  "self_revoke": {
                    "summary": "Cannot revoke own approval",
                    "value": {
                      "detail": "Admins cannot revoke their own approval."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "examples": {
                  "missing_token": {
                    "summary": "No token provided",
                    "value": {
                      "detail": "Not authenticated"
                    }
                  },
                  "invalid_token": {
                    "summary": "Invalid or expired JWT",
                    "value": {
                      "detail": "Could not validate credentials"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "examples": {
                  "admin_required": {
                    "summary": "Org admin privileges needed",
                    "value": {
                      "detail": "Organization admin privileges required."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Staff user not found",
            "content": {
              "application/json": {
                "examples": {
                  "staff_not_found": {
                    "summary": "Staff ID does not exist",
                    "value": {
                      "detail": "Staff user not found."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/staff/{user_id}/role": {
      "patch": {
        "tags": [
          "staff"
        ],
        "summary": "Promote or demote a staff member's role (site admin only)",
        "operationId": "set_role_staff__user_id__role_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleRequest"
              },
              "examples": {
                "promote_to_admin": {
                  "summary": "Promote to admin",
                  "description": "Give a staff member admin privileges",
                  "value": {
                    "role": "ADMIN"
                  }
                },
                "demote_to_staff": {
                  "summary": "Demote to staff",
                  "description": "Remove admin privileges from a user",
                  "value": {
                    "role": "STAFF"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                },
                "examples": {
                  "promoted": {
                    "summary": "User promoted to admin",
                    "value": {
                      "id": 2,
                      "username": "carlos_staff",
                      "is_approved": true,
                      "role": {
                        "id": 2,
                        "name": "ADMIN"
                      },
                      "created_at": "2024-03-20T18:30:00Z",
                      "is_disabled": false
                    }
                  },
                  "demoted": {
                    "summary": "User demoted to staff",
                    "value": {
                      "id": 2,
                      "username": "carlos_staff",
                      "is_approved": true,
                      "role": {
                        "id": 1,
                        "name": "USER"
                      },
                      "created_at": "2024-03-20T18:30:00Z",
                      "is_disabled": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid self-operation",
            "content": {
              "application/json": {
                "examples": {
                  "self_demote": {
                    "summary": "Cannot demote yourself",
                    "value": {
                      "detail": "Admins cannot demote themselves."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "examples": {
                  "missing_token": {
                    "summary": "No token provided",
                    "value": {
                      "detail": "Not authenticated"
                    }
                  },
                  "invalid_token": {
                    "summary": "Invalid or expired JWT",
                    "value": {
                      "detail": "Could not validate credentials"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "examples": {
                  "admin_required": {
                    "summary": "Org admin privileges needed",
                    "value": {
                      "detail": "Organization admin privileges required."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Staff user not found",
            "content": {
              "application/json": {
                "examples": {
                  "staff_not_found": {
                    "summary": "Staff ID does not exist",
                    "value": {
                      "detail": "Staff user not found."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/staff/{user_id}/disable": {
      "patch": {
        "tags": [
          "staff"
        ],
        "summary": "Disable a staff account (site admin only)",
        "description": "Disable a staff account (soft-delete).\n\nDisabled accounts cannot log in but their data is preserved for\naudit trails. Records who disabled the account, when, and why.",
        "operationId": "disable_staff_staff__user_id__disable_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisableUserRequest"
              },
              "examples": {
                "disable": {
                  "summary": "Disable with reason",
                  "description": "Disable a staff account. A reason is required for audit purposes.",
                  "value": {
                    "reason": "Left the organization and no longer volunteers"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                },
                "examples": {
                  "disabled": {
                    "summary": "Account now disabled",
                    "value": {
                      "id": 4,
                      "username": "ex_voluntario",
                      "is_approved": true,
                      "role": {
                        "id": 1,
                        "name": "USER"
                      },
                      "created_at": "2024-03-15T20:00:00Z",
                      "is_disabled": true,
                      "disabled_at": "2024-03-22T21:15:00Z",
                      "disabled_reason": "Left the organization"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid self-operation",
            "content": {
              "application/json": {
                "examples": {
                  "self_disable": {
                    "summary": "Cannot disable own account",
                    "value": {
                      "detail": "Admins cannot disable their own account."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing credentials",
            "content": {
              "application/json": {
                "examples": {
                  "missing_token": {
                    "summary": "No token provided",
                    "value": {
                      "detail": "Not authenticated"
                    }
                  },
                  "invalid_token": {
                    "summary": "Invalid or expired JWT",
                    "value": {
                      "detail": "Could not validate credentials"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "examples": {
                  "admin_required": {
                    "summary": "Org admin privileges needed",
                    "value": {
                      "detail": "Organization admin privileges required."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Staff user not found",
            "content": {
              "application/json": {
                "examples": {
                  "staff_not_found": {
                    "summary": "Staff ID does not exist",
                    "value": {
                      "detail": "Staff user not found."
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Account already disabled",
            "content": {
              "application/json": {
                "examples": {
                  "already_disabled": {
                    "summary": "Cannot disable twice",
                    "value": {
                      "detail": "This account is already disabled."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "examples": {
                  "reason_too_short": {
                    "summary": "Reason too short",
                    "value": {
                      "detail": [
                        {
                          "type": "string_too_short",
                          "loc": [
                            "body",
                            "reason"
                          ],
                          "msg": "String should have at least 5 characters",
                          "input": "bad",
                          "ctx": {
                            "min_length": 5
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/staff/{user_id}/recovery-code": {
      "post": {
        "tags": [
          "staff"
        ],
        "summary": "Generate a 6-digit password recovery code for a staff account (admin only)",
        "description": "Generate a one-time 6-digit recovery code for the given user.\n\nThe code expires after 72 hours and is invalidated once used.\nIf the user already had a code, it is overwritten (implicitly invalidated).\nThe admin is responsible for communicating the code to the user out-of-band.",
        "operationId": "generate_recovery_code_staff__user_id__recovery_code_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApproveUserRequest": {
        "properties": {
          "is_approved": {
            "type": "boolean",
            "title": "Is Approved"
          }
        },
        "type": "object",
        "required": [
          "is_approved"
        ],
        "title": "ApproveUserRequest",
        "description": "Request body for approving or revoking a user's access.\n\nAdmins use this to flip the is_approved flag on staff accounts.\nNew registrations start unapproved and require admin action."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "is_approved": {
            "type": "boolean",
            "title": "Is Approved"
          },
          "role": {
            "$ref": "#/components/schemas/RolePublic"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "is_disabled": {
            "type": "boolean",
            "title": "Is Disabled",
            "default": false
          },
          "disabled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disabled At"
          },
          "disabled_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disabled Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "is_approved",
          "role",
          "created_at"
        ],
        "title": "UserPublic",
        "description": "Public representation of a staff user.\n\nReturned when fetching user details. Includes the nested role object\nfor convenience. Never includes the hashed password."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/token"
          }
        }
      }
    }
  },
...
}